Changes in this commit include the following from teh FHWA feedbackk spreadsheet
15 - Tested on all states that we have data for. Successful.
22 - Added separate section for the ramp details
12 - Respecified the groupings of the results by Interstate, Non-Interstate NHS, Other/Minor Arterials and Collectors + Locals
2  - redesigned the histograms to reflect if we expect values to change year over year

Other changes include:
Implementing the labels for discretely coded variables
Respecified the import process to take advantage of sqldf for the merging. This reduces the memory footprint.
Added a bar plot for the discrete coded variables in the summary pages. This will replace the density plots for some variables.
removed section counts from the summary tables to save space.
Other misc changes.